+++ lib/src/grinder_context.dart
@@ -3,7 +3,6 @@
 
 library grinder.src.grinder_context;
 
-import 'ansi.dart' as ansi;
 import 'grinder.dart';
 import 'grinder_exception.dart';
 import 'grinder_task.dart';
@@ -38,7 +37,7 @@ class GrinderContext {
 
   /// Halt task execution; throws an exception with the given error message.
   void fail(String message) {
-    log('${ansi.bold}failed:${ansi.reset} ${ansi.red}${message}${ansi.reset}');
+    log(grinder.ansi.emphasized('failed: ') + grinder.ansi.error(message));
     throw new GrinderException(message);
   }