+++ tool/grind.dart
@@ -8,11 +8,11 @@ import 'package:grinder/grinder.dart';
 main(args) => grind(args);
 
 @Task()
-analyze() => new PubApp.global('tuneup')..runAsync(['check', '--ignore-infos']);
+analyze() => new PubApp.global('tuneup').runAsync(['check', '--ignore-infos']);
 
 @Task()
 test() {
-  //new TestRunner().testAsync();
+  // new TestRunner().testAsync();
   return Dart.runAsync(getFile('test/all.dart').path, vmArgs: ['--checked']);
 }